+++ src/components/ProviderSettings.vue
@@ -111,13 +111,12 @@
 
 <script>
 
-import { computed } from 'vue';
+import { computed, ref } from 'vue';
 import { storeToRefs } from "pinia";
 import { useSettingsStore } from '../stores/settings-store.js';
 import { useQuasar } from 'quasar';
 import { useI18n } from 'vue-i18n';
 import defaultProviders from '../services/providers.config.json';
-import { ref } from "vue";
 
 export default {
     setup() {